extract structure reader
