+++ src/cpp/benders/benders_core/include/antares-xpansion/benders/benders_core/CouplingMapGenerator.h
@@ -1,0 +1,6 @@
+#pragma once
+#include "common.h"
+
+struct CouplingMapGenerator {
+  static CouplingMap build_input(const std::filesystem::path &structure_path);
+};
