+++ src/cpp/benders/benders_core/include/antares-xpansion/benders/benders_core/common.h
@@ -191,7 +191,4 @@ struct BendersBaseOptions : public BaseOptions {
 };
 
 void usage(int argc);
-struct CouplingMapGenerator {
-  static CouplingMap build_input(const std::filesystem::path &structure_path);
-};
 Json::Value get_json_file_content(const std::filesystem::path &json_file);
