+++ src/cpp/benders/merge_mps/MergeMPS.cpp
@@ -2,8 +2,7 @@
 
 #include <filesystem>
 
-#include "antares-xpansion/helpers/ArchiveReader.h"
-#include "antares-xpansion/xpansion_interfaces/LogUtils.h"
+#include "antares-xpansion/benders/benders_core/CouplingMapGenerator.h"
 #include "antares-xpansion/helpers/Timer.h"
 
 MergeMPS::MergeMPS(const MergeMPSOptions &options, Logger &logger,
