+++ tests/cpp/outer_loop/outer_loop_test.cpp
@@ -1,4 +1,5 @@
 
+#include "antares-xpansion/benders/benders_core/CouplingMapGenerator.h"
 #include "antares-xpansion/benders/benders_core/CriterionInputDataReader.h"
 #include "antares-xpansion/benders/benders_core/MasterUpdate.h"
 #include "antares-xpansion/benders/benders_core/VariablesGroup.h"
